% Generated by roxygen2 (4.0.0): do not edit by hand
\name{add_error}
\alias{add_error}
\title{<brief desc>}
\usage{
add_error(tFrags, error_rate = 0.005)
}
\arguments{
\item{tFrags}{<what param does>}

\item{error_rate}{= 0.005 <what param does>}
}
\value{
fragments with error added
}
\description{
<full description>
}
\examples{
\dontrun{

}
}

