% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gc_bias.R
\name{add_gc_bias}
\alias{add_gc_bias}
\title{add GC bias to a count matrix}
\usage{
add_gc_bias(readmat, gcbias, transcripts)
}
\arguments{
\item{readmat}{matrix of counts, with rows corresponding to features
(transcripts) and columns corresponding to replicates}

\item{gcbias}{List of GC bias models to add to readmat. Must have length
equal to the number of columns of \code{readmat}. List elements must
either be integers 0 through 7, where 0 means no bias and 1-7 correspond
to built-in GC bias models, or objects of class \code{loess} which can
predict a deviation from overall mean count (on the log scale) given
a GC percentage between 0 and 1.}

\item{transcripts}{\code{DNAStringSet} object containing the sequences of
the features (transcripts) corresponding to the rows of \code{readmat}.
Length must be equal to the number of rows in \code{readmat}.}
}
\value{
matrix of the same size as \code{readmat}, but with counts for each
  replicate biased according to \code{gcbias}.
}
\description{
Given a matrix with rows corresponding to transcripts and sample-specific
GC bias models, bias the count matrix using the bias model.
}
\details{
Designed for internal use in \code{simulate_experiment} functions.
}
\examples{
  library(Biostrings)
  fastapath = system.file("extdata", "chr22.fa", package="polyester")
  numtx = count_transcripts(fastapath)
  transcripts = readDNAStringSet(fastapath)

  # create a count matrix:
  readmat = matrix(20, ncol=10, nrow=numtx)
  readmat[1:30, 1:5] = 40

  # add biases randomly: use built-in bias models
  set.seed(137)
  biases = sample(0:7, 10, replace=TRUE)
  readmat_biased = add_gc_bias(readmat, as.list(biases), transcripts)
}
