% Generated by roxygen2 (4.0.0): do not edit by hand
\name{count_transcripts}
\alias{count_transcripts}
\title{<brief desc>}
\usage{
count_transcripts(fasta)
}
\arguments{
\item{fasta}{<what param does>}
}
\value{
Number of transcripts annotated in \code{fasta}
}
\description{
<full description>
}
\examples{
\dontrun{

}
}

