% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_read_numbers}
\alias{create_read_numbers}
\title{Generate a simulated data set based on known model parameters}
\usage{
create_read_numbers(mu, fit, mod, beta, p0, seed = NULL)
}
\arguments{
\item{mu}{Baseline mean expression for negative binomial model}

\item{fit}{Fitted relationship between log mean and log size}

\item{mod}{Model matrix you would like to simulate from without an intercept}

\item{beta}{set of coefficients for the model matrix (must have same number of columns as mod)}

\item{p0}{A vector of the probabilities a count is zero}

\item{seed}{optional seed to set (for reproducibility)}
}
\value{
counts Data matrix with counts for genes in rows and samples in columns
}
\description{
Generate a simulated data set based on known model parameters
}
\author{
Jeff Leek
}

