% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_reads.R
\name{GenerateQ}
\alias{GenerateQ}
\title{Random quality generator}
\usage{
GenerateQ(len, num)
}
\arguments{
\item{len}{An \code{integer} represents the length of reads.}

\item{num}{An \code{integer} represents the reads number.}
}
\value{
A \code{BStringSet}
}
\description{
Generate quality
}
\details{
Randomly generate read qualities.
}
\author{
Yulong Niu \email{yulong.niu@hotmail.com}
}
\keyword{internal}
