% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_reads}
\alias{get_reads}
\title{get sequencing reads from fragments}
\usage{
get_reads(tFrags, readlen, paired = TRUE)
}
\arguments{
\item{tFrags}{DNAStringSet representing fragments}

\item{readlen}{Read length.}

\item{paired}{If \code{FALSE}, return only the first \code{readlen} bases of each element of
\code{tFrags} in the result; if \code{TRUE}, also return last \code{readlen} bases.}
}
\value{
DNAStringSet representing simulated RNA-seq reads
}
\description{
simulate the sequencing process by returning the sequence of one or both ends of provided
fragments
}
\seealso{
\code{\link{simulate_experiment}}, \code{\link{simulate_experiment_countmat}}
}

