% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getAttributeField}
\alias{getAttributeField}
\title{extract a specific field of the "attributes" column of a data frame created from a GTF/GFF file}
\usage{
getAttributeField(x, field, attrsep = "; ")
}
\arguments{
\item{x}{vector representing the "attributes" column of GTF/GFF file}

\item{field}{name of the field you want to extract from the "attributes" column}

\item{attrsep}{separator for the fields in the attributes column.  Defaults to '; ',
the separator for GTF files outputted by Cufflinks.}
}
\value{
vector of nucleotide positions included in the transcript
}
\description{
extract a specific field of the "attributes" column of a data frame created from a GTF/GFF file
}
\examples{
\dontrun{
# pre-loaded GTF file from chr22 (could also use gffRead in "ballgown" package):
data(gtf_dataframe)

# extract gene IDs from attributes column:
gtf_dataframe$gene_id = getAttributeField(gtf_dataframe$attributes, field="gene_id")
}
}
\author{
Wolfgang Huber, in the \code{davidTiling} package (LGPL license)
}
\seealso{
\url{http://useast.ensembl.org/info/website/upload/gff.html}, for specifics of the
GFF/GTF file format.
}

