% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{gtf_dataframe}
\alias{gtf_dataframe}
\title{data frame (in gtf-inspired format) for chromosome 22, hg19}
\format{data frame, 9 columns, 17769 rows}
\source{
Illumina iGenomes, hg19, 6 March 2013 version:
\url{http://ccb.jhu.edu/software/tophat/igenomes.shtml}.
}
\description{
In the data frame \code{gtf_dataframe}, each
  row corresponds to an exon / coding sequence / start codon / stop codon,
  and the columns correspond to standard GTF columns denoting annotated
  genomic features. See
  \url{http://www.ensembl.org/info/website/upload/gff.html}.
}

