% Generated by roxygen2 (4.0.0): do not edit by hand
\name{reverse_complement}
\alias{reverse_complement}
\title{reverse-complement some fragments}
\usage{
reverse_complement(tObj, seed = NULL)
}
\arguments{
\item{tObj}{DNAStringSet representing sequences.}

\item{seed}{optional seed to set before randomly selecting the sequences to be
reverse-complemented.}
}
\value{
DNAStringSet that is the same as \code{tObj}, but with about half the sequences
reverse-complemented.
}
\description{
randomly reverse-complement half of the sequences in a DNAStringSet
}

