% Generated by roxygen2 (4.0.0): do not edit by hand
\name{seq_gtf}
\alias{seq_gtf}
\title{Get transcript sequences from GTF file and sequence info}
\usage{
seq_gtf(gtf, seqpath, exononly = TRUE, idfield = "transcript_id",
  attrsep = "; ")
}
\arguments{
\item{gtf}{path to GTF file}

\item{seqpath}{path to folder containing one FASTA file (\code{.fa} extension) for each
chromosome in \code{gtf}}

\item{exononly}{if \code{TRUE} (as it is by default), only create transcript sequences from the
features labeled \code{exon} in \code{gtf}}

\item{idfield}{in the \code{attributes} column of \code{gtf}, what is the name of the field
identifying transcripts? Should be character. Default \code{"transcript_id"}.}

\item{attrsep}{in the \code{attributes} column of \code{gtf}, how are attributes separated?
Default \code{"; "}.}
}
\value{
DNAStringSet containing transcript sequences, with names corresponding to \code{idfield}
in \code{gtf}
}
\description{
Given a GTF file (for transcript structure) and DNA sequences, return a DNAStringSet
of transcript sequences
}
\references{
\url{http://www.ensembl.org/info/website/upload/gff.html}
}

