% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgseq.R
\name{sgseq}
\alias{sgseq}
\title{internal sequencing function
import doParallel}
\usage{
sgseq(readmat, transcripts, paired, outdir, extras, reportCoverage = FALSE,
  ncores = 2)
}
\description{
internal sequencing function
import doParallel
}

