% Generated by roxygen2 (4.0.0): do not edit by hand
\name{add_error}
\alias{add_error}
\title{add sequencing error to simulated reads}
\usage{
add_error(tFrags, error_rate = 0.005)
}
\arguments{
\item{tFrags}{DNAStringSet representing sequencing reads}

\item{error_rate}{error probability}
}
\value{
DNAStringSet equivalent to \code{tFrags} but with random sequencing errors inserted
}
\description{
simulate sequencing error by randomly changing the sequenced nucleotide on some of the reads
}

