% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_error.R
\name{add_error}
\alias{add_error}
\title{add sequencing error to simulated reads}
\usage{
add_error(tFrags, error_rate = 0.005)
}
\arguments{
\item{tFrags}{DNAStringSet representing sequencing reads}

\item{error_rate}{error probability}
}
\value{
DNAStringSet equivalent to \code{tFrags} but with random sequencing
  errors inserted
}
\description{
simulate sequencing error by randomly changing the sequenced nucleotide on
  some of the reads
}
\examples{
  library(Biostrings)
  data(srPhiX174)
  set.seed(174)
  srPhiX174_withError = add_error(srPhiX174)
  #error was introduced in, e.g., position 10 of 2nd string in set.
}
