% Generated by roxygen2 (4.0.0): do not edit by hand
\name{count_transcripts}
\alias{count_transcripts}
\title{determine how many transcripts are annotated in a FASTA file}
\usage{
count_transcripts(fasta)
}
\arguments{
\item{fasta}{character, path to a file in FASTA format}
}
\value{
Number of transcripts annotated in \code{fasta}
}
\description{
determine how many transcripts are annotated in a FASTA file
}
\examples{
fastapath = system.file("data", "chr22.fa", package="polyester")
count_transcripts(fastapath) #918
}

