% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reads.R
\name{get_reads}
\alias{get_reads}
\title{get sequencing reads from fragments}
\usage{
get_reads(tFrags, readlen, paired = TRUE)
}
\arguments{
\item{tFrags}{DNAStringSet representing fragments}

\item{readlen}{Read length.}

\item{paired}{If \code{FALSE}, return only the first \code{readlen} bases of 
each element of \code{tFrags} in the result; if \code{TRUE}, also return
last \code{readlen} bases.}
}
\value{
DNAStringSet representing simulated RNA-seq reads
}
\description{
simulate the sequencing process by returning the sequence of one or both ends of provided 
fragments
}
\examples{
  library(Biostrings)
  data(srPhiX174)
  set.seed(174)
  srPhiX174_reads = get_reads(srPhiX174, readlen=15, paired=FALSE)
  srPhiX174_reads  
  # set of single-end, 15bp reads, treating srPhiX174 as the fragments
}
\seealso{
\code{\link{simulate_experiment}}, 
  \code{\link{simulate_experiment_countmat}}
}
