% Generated by roxygen2 (4.0.0): do not edit by hand
\name{NB}
\alias{NB}
\title{Negative binomial draws}
\usage{
NB(basemeans, dispersion_param, seed = NULL)
}
\arguments{
\item{basemeans}{<what param does>}

\item{dispersion_param}{<what param does>}

\item{seed=NULL}{<what param does>}
}
\description{
<full description>
}
\examples{
\dontrun{

}
}

