% Generated by roxygen2 (4.0.0): do not edit by hand
\name{simulate_experiment}
\alias{simulate_experiment}
\title{simulate RNA-seq experiment using negative binomial model}
\usage{
simulate_experiment(fasta = NULL, gtf = NULL, seqpath = NULL,
  num_reps = 10, fraglen = 250, fragsd = 25, readlen = 100,
  error_rate = 0.005, paired = TRUE, reads_per_transcript = 300,
  fold_changes, dispersion_param = NULL, outdir = "", write_info = TRUE,
  transcriptid = NULL, seed = NULL, ...)
}
\arguments{
\item{fasta}{path to FASTA file containing transcripts from which to simulate reads. See details.}

\item{gtf}{path to GTF file containing transcript structures from which reads should be
simulated. See details.}

\item{seqpath}{path to folder containing one FASTA file (\code{.fa} extension) for each
chromosome in \code{gtf}. See details.}

\item{num_reps}{How many biological replicates should be in each group? If \code{num_reps}
is an integer, \code{num_reps} replicates will be simulated in each group.
Otherwise, \code{num_reps} can be a length-2 vector, where \code{num_reps[1]}
and \code{num_reps[2]} replicates will be simulated in each of the two groups.}

\item{fraglen}{Mean RNA fragment length. Sequences will be read off the end(s) of these
fragments.}

\item{fragsd}{Standard deviation of fragment lengths.}

\item{readlen}{Read length.}

\item{error_rate}{Sequencing error rate. Must be between 0 and 1. A uniform error model is
assumed.}

\item{paired}{If \code{TRUE}, paired-end reads are simulated; else single-end reads are
simulated.}

\item{reads_per_transcript}{baseline mean number of reads to simulate
from each transcript. Can be an integer, in which case this many reads
are simulated from each transcript, or an integer vector whose length
matches the number of transcripts in \code{fasta}.}

\item{fold_changes}{Vector of multiplicative fold changes between groups,
one entry per transcript in \code{fasta}. A fold change > 1 means the
transcript is overexpressed in the first \code{num_reps} (or \code{num_reps[1]})
samples. Fold change < 1 means transcript is overexpressed in the last
\code{num_reps} (or \code{num_reps[2]}) samples. The change is in the mean
number of reads generated from the transcript, between groups.}

\item{dispersion_param}{the negative binomial \code{size} distribution
(see \code{\link{NegBinomial}})
of the number of reads drawn per transcript. If left blank, defaults to
\code{reads_per_transcript / 3}. Negative binomial variance is mean + mean^2 / size.}

\item{outdir}{character, path to folder where simulated reads should be written. Should end
with "/" if specified. If unspecified, reads are written to current working directory.}

\item{write_info}{If \code{TRUE}, write a file matching transcript IDs to differential expression
status into the file \code{outdir/sim_info.txt}.}

\item{transcriptid}{optional vector of transcript IDs to be written into \code{sim_info.txt}.
Defaults to \code{names(readDNAStringSet(fasta))}. This option is useful if default names are
very long or contain special characters.}

\item{seed}{Optional seed to set before simulating reads, for reproducibility.}

\item{...}{additional arguments to pass to \code{seq_gtf} if using \code{gtf} and \code{seqpath}}
}
\description{
create FASTA files containing RNA-seq reads simulated from provided transcripts, with optional
differential expression between two groups
}
\details{
Reads can either be simulated from a FASTA file of transcripts (provided with the
\code{fasta} argument) or from a GTF file plus DNA sequences (provided with the \code{gtf} and
\code{seqpath} arguments). Simulating from a GTF file and DNA sequences may be a bit slower:
it took about 6 minutes to parse the GTF/sequence files for chromosomes 1-22, X, and Y in hg19.
}
\examples{
## simulate a few reads from chromosome 22

fastapath = system.file("data", "chr22.fa", package="polyester")
numtx = count_transcripts(fastapath)
set.seed(4)
fold_changes = sample(c(0.5, 1, 2), size=numtx, prob=c(0.05, 0.9, 0.05), replace=TRUE)
require(Biostrings)
tNames = gsub("'", "", names(readDNAStringSet(fastapath))) #remove quotes in transcript IDs

simulate_experiment(fastapath, reads_per_transcript=10, fold_changes=fold_changes,
    outdir="./simdata/", transcriptid=tNames, seed=12)
}

