% Generated by roxygen2 (4.0.0): do not edit by hand
\name{simulate_experiment_countmat}
\alias{simulate_experiment_countmat}
\title{<brief desc>}
\usage{
simulate_experiment_countmat(fasta, readmat, outdir, fraglen = 250,
  fragsd = 25, readlen = 100, error_rate = 0.005, paired = TRUE)
}
\arguments{
\item{fasta}{<what param does>}

\item{readmat}{<what param does>}

\item{outdir}{<what param does>}

\item{fraglen=250}{<what param does>}

\item{fragsd=25}{<what param does>}

\item{readlen=100}{<what param does>}

\item{error_rate=0.005}{<what param does>}

\item{paired=TRUE}{<what param does>}
}
\description{
<full description>
}
\examples{
\dontrun{

}
}

